Energy Aspects - Assessment/requirements.txt
openpyxl
pandas
psycopg2-binary
pycountry
spacy
sqlalchemy